create common component
