+++ web/src/App.jsx
@@ -14,8 +14,8 @@ import Register from './pages/Register';
 import { checkUserSession } from './contexts/AuthActions';
 import LinearProgressModal from './components/common/LinearProgressModal';
 import Home from './pages/Home';
-import Posts from './pages/Posts';
 import PostsByTopics from './pages/PostsByTopics';
+import PostsByOwner from './pages/PostsByOwner';
 
 const theme = createTheme(({
   palette: {
@@ -96,7 +96,7 @@ function Wrapper() {
             path={`/${FE_ROUTES.POSTS}/:type`}
             element={(
               <AuthRouteComponent
-                Success={<Dashboard><Posts /></Dashboard>}
+                Success={<Dashboard><PostsByOwner /></Dashboard>}
                 Terminated={<Navigate to={`/${FE_ROUTES.LOGIN}`} />}
               />
             )}
