+++ web/src/components/common/ContentWrapper.jsx
@@ -1,23 +1,35 @@
 import { styled } from '@mui/material/styles';
+import { topbarHeight } from '../../constants/styles';
+import { GRAY3, WHITE } from '../../constants/colors';
 
-const PREFIX = 'ContentWrapper';
-
-const classes = {
-  wrapper: `${PREFIX}-wrapper`,
-};
-
-const Root = styled('div')(({ theme }) => ({
-  [`&.${classes.wrapper}`]: {
-    padding: theme.spacing(0, 2),
-  },
+const ContentWrapperContainer = styled('div')(({ theme }) => ({
+  padding: theme.spacing(0, 2),
 }));
 
-function ContentWrapper({ Header, children }) {
+const ContentWrapperHeaderContainer = styled('div')(({ theme }) => ({
+  position: 'sticky',
+  top: topbarHeight,
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+  padding: theme.spacing(1, 2),
+  margin: theme.spacing(2, 0),
+  border: `1px solid ${GRAY3}`,
+  borderRadius: 4,
+  backgroundColor: WHITE,
+  zIndex: 100,
+}));
+
+function ContentWrapper({ LeftHeader, RightHeader, children }) {
   return (
-    <Root className={classes.wrapper}>
-      {Header}
+    <ContentWrapperContainer>
+      <ContentWrapperHeaderContainer>
+        {LeftHeader}
+        {RightHeader}
+      </ContentWrapperHeaderContainer>
       {children}
-    </Root>
+    </ContentWrapperContainer>
   );
 }
